fix: add Windows path conversion to lisp_tools and fix 32-bit compilation
- Add resolvePath() helper to lisp_tools.go that calls PosixToWindowsPath
  on Windows, applied to all path parameters in file operation handlers
- Fix 32-bit compilation overflow errors by adding explicit type casts
  for large integer constants in ffi_math.go, ffi_hash_crc32.go,
  ffi_math_big.go, ffi_debug_pe.go, and ffi_text_scanner.go

+++ microlisp/ffi_hash_crc32.go
@@ -19,9 +19,9 @@ func register_hash_crc32() {
 			"NewIEEE": crc32.NewIEEE,
 			"Update": crc32.Update,
 			"IEEETable": crc32.IEEETable,
-			"Castagnoli": crc32.Castagnoli,
-			"IEEE": crc32.IEEE,
-			"Koopman": crc32.Koopman,
+			"Castagnoli": uint32(crc32.Castagnoli),
+			"IEEE": uint32(crc32.IEEE),
+			"Koopman": uint32(crc32.Koopman),
 			"Size": crc32.Size,
 		},
 	)
